@@ -2,9 +2,10 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-	<TargetFramework>net4.6-windows</TargetFramework>
+	<TargetFramework>net6.0-windows</TargetFramework>
 	<RootNamespace>Tetris_NEAT_AI</RootNamespace>
 	<AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+    <ImplicitUsings>true</ImplicitUsings>
 	<UseWPF>true</UseWPF>
 	<AssemblyName>Budget Tetris</AssemblyName>
   </PropertyGroup>
@@ -86,6 +87,12 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.XAudio2.Redist" Version="1.2.10" />
     <PackageReference Include="System.ValueTuple" Version="4.5.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Reference Include="Bass.Net">
+      <HintPath>..\Bass24.Net\full\Bass.Net.dll</HintPath>
+    </Reference>
   </ItemGroup>
 
 </Project>
