@@ -1,47 +1,17 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Threading;
-using System.Windows.Media;
-using System.Windows.Threading;
+﻿using System.IO;
 
 namespace Tetris
 {
     class Sound
     {
         public const string SoundsFolder = @"Sounds\";
-        public const int MaxSounds = 16;
 
+        public static bool Initialised { get; private set; } = false;
         public static Thread SoundThread { get; private set; }
-        public static Dispatcher SoundDP { get => Dispatcher.FromThread(SoundThread); }
 
-        private static bool _IsMuted = false;
-        public static bool IsMuted
-        {
-            get => _IsMuted;
-            set
-            {
-                _IsMuted = value;
-                SoundDP?.Invoke(() => BGMPlayer.IsMuted = value);
-            }
-        }
-
-        public static double SFXVolume;
-        private static readonly HashSet<MediaPlayer> MediaPlayers = new HashSet<MediaPlayer>();
-        private static readonly Queue<MediaPlayer> MediaPlayerPool = new Queue<MediaPlayer>();
-
-        private static double _BGMVolume;
-        public static double BGMVolume
-        {
-            get => _BGMVolume;
-            set
-            {
-                _BGMVolume = value;
-                if (BGMPlayer != null)
-                    SoundDP.Invoke(() => BGMPlayer.Volume = value);
-            }
-        }
-        private static MediaPlayer BGMPlayer;
+        public static bool IsMuted { get; set; }
+        public static double SFXVolume { get; set; }
+        public static double BGMVolume { get; set; }
 
         public static readonly Sound BGM = new Sound("Korobeiniki Remix.wav"),
                                      SoftDrop = new Sound("bfall.wav"),
@@ -64,80 +34,24 @@ namespace Tetris
         };
 
 
-        public Uri Path { get; private set; }
-        public bool HasSource;
+        private readonly int Handle = 0;
+        public bool Loaded { get => Handle != 0; }
 
         public Sound(string file_name)
         {
-            Path = new Uri(SoundsFolder + file_name, UriKind.Relative);
-            HasSource = File.Exists(Path.ToString());
+            if (!Initialised) return;
         }
 
         public static void InitSound()
         {
-            if (SoundThread != null) return;
-
-            // Create Sound thread
-            SoundThread = new Thread(() =>
-            {
-                BGMPlayer = new MediaPlayer
-                {
-                    Volume = BGMVolume,
-                };
-                // Loop delegate
-                BGMPlayer.MediaEnded += (s, e) =>
-                {
-                    BGMPlayer.Position = TimeSpan.Zero;
-                    BGMPlayer.Play();
-                };
-                // Play BGM looping
-                BGMPlayer.Open(BGM.Path);
-                BGMPlayer.Play();
-                // Run the dispatcher
-                Dispatcher.Run();
-            });
-            SoundThread.Start();
-            SoundThread.Priority = ThreadPriority.Lowest;
+            return;
         }
 
         public void Play()
         {
-            if (!HasSource || IsMuted) return;
+            return;
 
-            SoundDP.Invoke(() =>
-            {
-                MediaPlayer player = TryGetSFXPlayer();
-                if (player == null) return;
-
-                player.Volume = SFXVolume;
-                player.Open(Path);
-                player.Play();
-            });
-        }
-
-        private static MediaPlayer TryGetSFXPlayer()
-        {
-            // Create a new player if pool is dry
-            if (MediaPlayerPool.Count == 0)
-            {
-                if (MediaPlayers.Count >= MaxSounds)
-                    return null;
-
-                MediaPlayer player = new MediaPlayer()
-                {
-                    Volume = SFXVolume,
-                };
-                MediaPlayers.Add(player);
-                player.MediaEnded += (s, e) =>
-                {
-                    player.Close();
-                    MediaPlayerPool.Enqueue(player);
-                };
-
-                return player;
-            }
-
-            return MediaPlayerPool.Dequeue();
+            //if (!Initialised || IsMuted || !Loaded) return;
         }
     }
 }
